refactor(feedback): move into sub-components
